One-for-All-125 - Moved the headerItems to server.
